+++ src/lib.rs
@@ -2,32 +2,45 @@
 pub struct QueryBuilder {
     pub query: String,
     pub table: String,
-    pub qtype: QueryType
+    pub qtype: QueryType,
+    pub list: Vec<KeywordList>
 }
 
 impl QueryBuilder {
     pub fn select(fields: Vec<&str>) -> std::result::Result<Self, std::io::Error> {
         match QueryBuilder::sanitize(fields.clone()) {
             Ok(_) => {
-                let mut query = "SELECT ".to_string();
+                if fields[0] == "*" {
+                    let query = "SELECT * FROM".to_string();
+    
+                    return Ok(QueryBuilder {
+                        query,
+                        table: "".to_string(),
+                        qtype: QueryType::Select,
+                        list: vec![KeywordList::Select]
+                    })
+                } else {
+                    let mut query = "SELECT ".to_string();
 
-                let length_of_fields = fields.len();
-
-                for (i , field) in fields.into_iter().enumerate() {
-                    if i + 1 == length_of_fields {
-                        query = format!("{}{} ", query, field);
-                    } else {
-                        query = format!("{}{}, ", query, field);
+                    let length_of_fields = fields.len();
+    
+                    for (i , field) in fields.into_iter().enumerate() {
+                        if i + 1 == length_of_fields {
+                            query = format!("{}{} ", query, field);
+                        } else {
+                            query = format!("{}{}, ", query, field);
+                        }
                     }
+    
+                    let query = format!("{}FROM", query);
+    
+                    return Ok(QueryBuilder {
+                        query,
+                        table: "".to_string(),
+                        qtype: QueryType::Select,
+                        list: vec![KeywordList::Select]
+                    })
                 }
-
-                let query = format!("{}FROM", query);
-
-                return Ok(QueryBuilder {
-                    query,
-                    table: "".to_string(),
-                    qtype: QueryType::Select
-                })
             },
             Err(_) => {
                 return Err(std::io::Error::new(std::io::ErrorKind::PermissionDenied, "query cannot build because inserted arbitrary query."))
@@ -40,6 +53,7 @@ impl QueryBuilder {
             query: "DELETE FROM".to_string(),
             table: "".to_string(),
             qtype: QueryType::Delete,
+            list: vec![KeywordList::Delete]
         })
     }
 
@@ -47,7 +61,8 @@ impl QueryBuilder {
         return Ok(QueryBuilder {
             query: "UPDATE".to_string(),
             table: "".to_string(),
-            qtype: QueryType::Delete,
+            qtype: QueryType::Update,
+            list: vec![KeywordList::Update]
         })
     }
 
@@ -126,11 +141,13 @@ impl QueryBuilder {
         }
 
         query = format!("{} {} VALUES {}", query, columns_string, values_string);
+        
 
         return Ok(Self {
             query,
             table: "".to_string(),
-            qtype: QueryType::Insert
+            qtype: QueryType::Insert,
+            list: vec![KeywordList::Insert]
         })
     }
 
@@ -154,13 +171,17 @@ impl QueryBuilder {
                 self.query = format!("{} {}", self.query, table);
                 self.table = table.to_string()
             },
-            QueryType::Null => panic!("You cannot add a table before you start a query")
+            QueryType::Null => panic!("You cannot add a table before you start a query"),
+            QueryType::Create => panic!("You cannot use create keyword with a QueryBuilder instance")
         }
+
+        self.list.push(KeywordList::Table);
 
         Self {
             query: self.query.clone(),
             table: self.table.clone(),
             qtype: self.qtype.clone(),
+            list: self.list.clone()
         }
     }
 
@@ -190,19 +211,23 @@ impl QueryBuilder {
                     }
                 }
 
+                self.list.push(KeywordList::Where);
                 return Self {
                     query: self.query.clone(),
                     table: self.table.clone(),
                     qtype: self.qtype.clone(),
+                    list: self.list.clone()
                 }
             },
             Err(error) => {
                 println!("An Error Occured when executing where method: {}", error);
 
+                self.list.push(KeywordList::Where);
                 return Self {
                     query: self.query.clone(),
                     table: self.table.clone(),
                     qtype: self.qtype.clone(),
+                    list: self.list.clone()
                 }
             }
         }
@@ -226,19 +251,24 @@ impl QueryBuilder {
                     }
                 }
 
+                self.list.push(KeywordList::Or);
+
                 return Self {
                     query: self.query.clone(),
                     table: self.table.clone(),
                     qtype: self.qtype.clone(),
+                    list: self.list.clone()
                 }
             },
             Err(error) => {
                 println!("An Error Occured when executing or method: {}", error);
 
+                self.list.push(KeywordList::Or);
                 return Self {
                     query: self.query.clone(),
                     table: self.table.clone(),
-                    qtype: self.qtype.clone()
+                    qtype: self.qtype.clone(),
+                    list: self.list.clone()
                 }
             }
         }
@@ -288,19 +318,23 @@ impl QueryBuilder {
                     }
                 }
 
+                self.list.push(KeywordList::Set);
                 return Self {
                     query: self.query.clone(),
                     table: self.table.clone(),
-                    qtype: self.qtype.clone()
+                    qtype: self.qtype.clone(),
+                    list: self.list.clone()
                 }
             },
             Err(error) => {
                 println!("An Error Occured when executing or method: {}", error);
 
+                self.list.push(KeywordList::Set);
                 return Self {
                     query: self.query.clone(),
                     table: self.table.clone(),
-                    qtype: self.qtype.clone()
+                    qtype: self.qtype.clone(),
+                    list: self.list.clone()
                 }
             }
         }
@@ -324,19 +358,23 @@ impl QueryBuilder {
                     }
                 }
 
+                self.list.push(KeywordList::And);
                 return Self {
                     query: self.query.clone(),
                     table: self.table.clone(),
-                    qtype: self.qtype.clone()
+                    qtype: self.qtype.clone(),
+                    list: self.list.clone()
                 }
             },
             Err(error) => {
                 println!("An Error Occured when executing and method: {}", error);
 
+                self.list.push(KeywordList::And);
                 return Self {
                     query: self.query.clone(),
                     table: self.table.clone(),
-                    qtype: self.qtype.clone()
+                    qtype: self.qtype.clone(),
+                    list: self.list.clone()
                 }
             }
         }
@@ -345,20 +383,24 @@ impl QueryBuilder {
     pub fn offset(&mut self, offset: i32) -> Self {
         self.query = format!("{} OFFSET {}", self.query, offset);
 
+        self.list.push(KeywordList::Offset);
         Self {
             query: self.query.clone(),
             table: self.table.clone(),
-            qtype: self.qtype.clone()
+            qtype: self.qtype.clone(),
+            list: self.list.clone()
         }
     }
 
     pub fn limit(&mut self, limit: i32) -> Self {
         self.query = format!("{} LIMIT {}", self.query, limit);
 
+        self.list.push(KeywordList::Limit);
         Self {
             query: self.query.clone(),
             table: self.table.clone(),
-            qtype: self.qtype.clone()
+            qtype: self.qtype.clone(),
+            list: self.list.clone()
         }
     }
 
@@ -370,10 +412,12 @@ impl QueryBuilder {
                     Err(error) => {
                         println!("That Error Occured in like method: {}", error);
         
+                        self.list.push(KeywordList::Like);
                         return Self {
                             query: self.query.clone(),
                             table: self.table.clone(),
-                            qtype: self.qtype.clone()
+                            qtype: self.qtype.clone(),
+                            list: self.list.clone()
                         }
                     }
                 }
@@ -386,19 +430,23 @@ impl QueryBuilder {
                     }
                 }
 
+                self.list.push(KeywordList::Like);
                 Self {
                     query: self.query.clone(),
                     table: self.table.clone(),
-                    qtype: self.qtype.clone()
+                    qtype: self.qtype.clone(),
+                    list: self.list.clone()
                 }
             },
             Err(error) => {
                 println!("That Error Occured in like method: {}", error);
 
+                self.list.push(KeywordList::Like);
                 Self {
                     query: self.query.clone(),
                     table: self.table.clone(),
-                    qtype: self.qtype.clone()
+                    qtype: self.qtype.clone(),
+                    list: self.list.clone()
                 }
             }
         }
@@ -414,10 +462,12 @@ impl QueryBuilder {
             Err(error) => {
                 println!("{}", error);
 
+                self.list.push(KeywordList::OrderBy);
                 return Self {
                     query: self.query.clone(),
                     table: self.table.clone(),
-                    qtype: self.qtype.clone()
+                    qtype: self.qtype.clone(),
+                    list: self.list.clone()
                 }
             }
         }
@@ -431,11 +481,12 @@ impl QueryBuilder {
         }
 
         self.query = format!("{} ORDER BY {} {}", self.query, column, ordering);
-
+        self.list.push(KeywordList::OrderBy);
         Self {
             query: self.query.clone(),
             table: self.table.clone(),
-            qtype: self.qtype.clone()
+            qtype: self.qtype.clone(),
+            list: self.list.clone()
         }
     }
 
@@ -445,15 +496,18 @@ impl QueryBuilder {
         }
 
         self.query = format!("{} ORDER BY RAND()", self.query);
-
+        self.list.push(KeywordList::OrderBy);
         Self {
             query: self.query.clone(),
             table: self.table.clone(),
-            qtype: self.qtype.clone()
+            qtype: self.qtype.clone(),
+            list: self.list.clone()
         }
     }
 
     pub fn finish(&mut self) -> String {
+        self.list.push(KeywordList::Finish);
+
         return format!("{};", self.query);
     }
 
@@ -463,35 +517,34 @@ impl QueryBuilder {
         };
 
         for field in fields.into_iter() {
-            if (field.contains(";") && (field.len() < 40)) ||
-               field.contains("; DROP") ||
-               field.contains("admin' #") ||
-               field.contains("admin'/*") ||
-               field.contains("; UNION") ||
-               field.contains("OR 1 = 1") ||
-               field.contains("OR 1 = 1#") ||
-               field.contains("OR 1 = 1/*") ||
-               field.contains("OR true = true") ||
-               field.contains("OR false = false") ||
-               field.contains("OR '1' = '1'") ||
-               field.contains("OR '1' = '1'#") ||
-               field.contains("OR '1' = '1'/*") ||
-               field.contains("; SLEEP(") ||
-               field.contains("; Sleep(") ||
-               field.contains("; sleep(") ||
-               field.contains("--") ||
-               field.contains("DROP TABLE") ||
-               field.contains("DROP SCHEMA") ||
-               field.contains("SELECT IF") ||
-               field.contains("UNION SELECT") ||
-               field.contains("UNION ALL") ||
-               field.contains("EXEC") ||
-               field.contains("master..") ||
-               field.contains("masters..") ||
-               field.contains("WAITFOR") ||
-               field.contains("DELAY") {
-                return Err(std::io::Error::new(std::io::ErrorKind::PermissionDenied, "You cannot run arbitrary queries"))
-               }
+            if (field.to_lowercase().contains(";") && (field.len() < 40)) ||
+                field.to_lowercase().contains("; drop") ||
+                field.to_lowercase().contains("admin' #") ||
+                field.to_lowercase().contains("admin'/*") ||
+                field.to_lowercase().contains("; union") ||
+                field.to_lowercase().contains("or 1 = 1") ||
+                field.to_lowercase().contains("or 1 = 1#") ||
+                field.to_lowercase().contains("or 1 = 1/*") ||
+                field.to_lowercase().contains("or true = true") ||
+                field.to_lowercase().contains("or false = false") ||
+                field.to_lowercase().contains("or '1' = '1'") ||
+                field.to_lowercase().contains("or '1' = '1'#") ||
+                field.to_lowercase().contains("or '1' = '1'/*") ||
+                field.to_lowercase().contains("; sleep(") ||
+                field.to_lowercase().contains("--") ||
+                field.to_lowercase().contains("drop table") ||
+                field.to_lowercase().contains("drop schema") ||
+                field.to_lowercase().contains("select if") ||
+                field.to_lowercase().contains("union select") ||
+                field.to_lowercase().contains("union all") ||
+                field.to_lowercase().contains("exec") ||
+                field.to_lowercase().contains("master..") ||
+                field.to_lowercase().contains("masters..") ||
+                field.to_lowercase().contains("information_schema") ||
+                field.to_lowercase().contains("load_file") ||
+                field.to_lowercase().contains("alter user") {
+                    return Err(std::io::Error::new(std::io::ErrorKind::PermissionDenied, "You cannot run arbitrary queries"))
+            }
         }
 
         return Ok("".to_string())
@@ -499,8 +552,119 @@ impl QueryBuilder {
 }
 
 #[derive(Debug, Clone)]
+pub struct SchemaBuilder {
+    pub query: String,
+    pub schema: String,
+    pub list: Vec<KeywordList>
+}
+
+impl SchemaBuilder {
+    pub fn create(name: &str) -> std::result::Result<Self, std::io::Error> {
+        if name.contains("!") ||
+           name.contains("-") ||
+           name.contains("=") ||
+           name.contains("+") ||
+           name.contains("%") ||
+           name.contains("$") ||
+           name.contains("&") ||
+           name.contains("#") ||
+           name.contains("[") ||
+           name.contains("]") ||
+           name.contains("{") ||
+           name.contains("}") ||
+           name.contains(":") ||
+           name.contains(";") ||
+           name.contains("'") ||
+           name.contains("\"") ||
+           name.contains(",") ||
+           name.contains(".") {
+                return Err(std::io::Error::new(std::io::ErrorKind::InvalidInput, "Schame name cannot include these characters: '!', '-', '=', '+', '%', '$', '&', '#', '[', ']', '{', '}', ':', ';', '\"', \"'\", '.', ','"))
+        }
+
+        Ok(Self {
+            query: format!("CREATE DATABASE {}", name),
+            schema: name.to_string(),
+            list: vec![KeywordList::Create]
+        })
+    }
+
+    pub fn use_another_schema(name: &str) -> std::result::Result<Self, std::io::Error> {
+        if name.contains("!") ||
+        name.contains("-") ||
+        name.contains("=") ||
+        name.contains("+") ||
+        name.contains("%") ||
+        name.contains("$") ||
+        name.contains("&") ||
+        name.contains("#") ||
+        name.contains("[") ||
+        name.contains("]") ||
+        name.contains("{") ||
+        name.contains("}") ||
+        name.contains(":") ||
+        name.contains(";") ||
+        name.contains("'") ||
+        name.contains("\"") ||
+        name.contains(",") ||
+        name.contains(".") {
+             return Err(std::io::Error::new(std::io::ErrorKind::InvalidInput, "Schame name cannot include these characters: '!', '-', '=', '+', '%', '$', '&', '#', '[', ']', '{', '}', ':', ';', '\"', \"'\", '.', ','"))
+        }
+
+        Ok(Self {
+            query: format!("USE {}", name),
+            schema: name.to_string(),
+            list: vec![KeywordList::Use, KeywordList::Create]
+        })
+    }
+
+    pub fn if_not_exists(&mut self) -> Self {
+        match self.list[0] {
+            KeywordList::Create => (),
+            KeywordList::Table => (),
+            _ => panic!("if_not_exists method cannot be used without Create or Table queries")
+        }
+
+        let split_the_query =  self.query.split(" DATABASE ").collect::<Vec<&str>>();
+        self.query = format!("{} DATABASE IF NOT EXISTS {}", split_the_query[0], split_the_query[1]);
+
+        self.list.insert(0, KeywordList::IfNotExist);
+        Self {
+            query: self.query.clone(),
+            schema: self.schema.clone(),
+            list: self.list.clone()
+        }
+    }
+
+    pub fn use_schema(&mut self, name: Option<&str>) -> Self {
+        match name {
+            Some(schema_name) => {
+                self.query = format!("USE {}", schema_name)
+            },
+            None => {
+                self.query = format!("USE {}", self.schema);
+            }
+        }
+
+        Self {
+            query: self.query.clone(),
+            schema: self.schema.clone(),
+            list: self.list.clone()
+        }
+    }
+
+    pub fn finish(&self) -> String {
+        return format!("{};", self.query)
+    }
+}
+
+#[derive(Debug, Clone)]
+pub enum KeywordList {
+    Select, Update, Delete, Insert, Table, Where, Or, And, Set, Finish, OrderBy, Like, Limit, Offset, IfNotExist, Create, Use 
+}
+
+#[derive(Debug, Clone)]
 pub enum QueryType {
-    Select, Update, Delete, Insert, Null
+    Select, Update, Delete, Insert, Null, Create
 }
 
 #[derive(Debug, Clone)]
@@ -511,6 +675,29 @@ pub enum ValueType {
 #[cfg(test)]
 mod test {
     use super::*;
+
+    #[test]
+    pub fn test_schema_query_declarative(){
+        let schema = SchemaBuilder::create("blog_website").unwrap().if_not_exists().finish();
+
+        assert_eq!("CREATE DATABASE IF NOT EXISTS blog_website;", schema);
+    }
+
+    #[test]
+    pub fn test_schema_query_imperative(){
+        let mut schema = SchemaBuilder::create("blog_website").unwrap();
+        schema.if_not_exists();
+        let schema_query = schema.finish();
+
+        assert_eq!("CREATE DATABASE IF NOT EXISTS blog_website;", schema_query);
+    }
+
+    #[test]
+    pub fn test_use_another_schema(){
+        let schema = SchemaBuilder::use_another_schema("chat_website").unwrap().finish();
+
+        assert_eq!("USE chat_website;", schema);
+    }
 
     #[test]
     pub fn test_insert_query(){
@@ -529,5 +716,38 @@ mod test {
         let update_query = QueryBuilder::update().unwrap().table("blogs").set("title", ("Hello Rust!", ValueType::String)).set("author", ("Necdet", ValueType::String)).finish();
 
         assert_eq!("UPDATE blogs SET title = 'Hello Rust!', author = 'Necdet';", update_query);
+    }
+
+    #[test]
+    pub fn test_delete_query(){
+        let delete_query = QueryBuilder::delete().unwrap().table("blogs").where_cond("id", "=", ("1", ValueType::String)).finish();
+
+        assert_eq!("DELETE FROM blogs WHERE id = '1';", delete_query);
+    }
+
+    #[test]
+    pub fn test_select_query_declarative(){
+        let mut select = QueryBuilder::select(["id", "title", "description", "point"].to_vec()).unwrap();
+
+        let select_query = select.table("blogs")
+                                    .where_cond("id", "=", ("10", ValueType::Integer))
+                                    .and("point", ">", ("90", ValueType::Integer))
+                                    .or("id", "=", ("20", ValueType::Integer))
+                                    .finish();
+
+        assert_eq!("SELECT id, title, description, point FROM blogs WHERE id = 10 AND point > 90 OR id = 20;", select_query)
+    }
+
+    #[test]
+    pub fn test_select_query_imperative(){
+        let mut select = QueryBuilder::select(["*"].to_vec()).unwrap();
+
+        let mut select_query = select.table("blogs");
+        select_query.where_cond("id", "=", ("5", ValueType::Integer));
+        select_query.or("id", "=", ("25", ValueType::Integer));
+
+        let finish_the_select_query = select_query.finish();
+
+        assert_eq!("SELECT * FROM blogs WHERE id = 5 OR id = 25;", finish_the_select_query);
     }
 }